Remove ConnectionManager class
Removed the ConnectionManager class due to errors encountered while attempting to write tests for it.  The pattern of a collection holding different Connection<T> objects does not work.  When attempting to add a Connection<Derived> to the collection and doing an "as" cast to the base, it returns null.  

I would put forth that the primary use case will have a single connection.  If the user needs to hold a collection of connections then they can manage that.

If we want to have a hetergenous collection manager then a new pattern will need to be done using generic interfaces (I think).

see
https://docs.microsoft.com/en-us/dotnet/standard/generics/covariance-and-contravariance